+++ inc/core/actions.php
@@ -36,7 +36,7 @@ function extrachill_default_archive_header() {
 add_action( 'extrachill_archive_header', 'extrachill_default_archive_header', 10 );
 
 function extrachill_default_search_header() {
-    include get_template_directory() . '/inc/archives/search/search-header.php';
+    include get_template_directory() . '/inc/archives/archive-header.php';
 }
 add_action( 'extrachill_search_header', 'extrachill_default_search_header', 10 );